ReviewItem and Commenting process : Fix #106 #100 #107 #108 #132

+++ UI_DSM/UI_DSM.Server/Modules/AnnotationModule.cs
@@ -17,6 +17,7 @@ namespace UI_DSM.Server.Modules
     using Microsoft.AspNetCore.Components;
 
     using UI_DSM.Server.Managers;
+    using UI_DSM.Server.Managers.AnnotatableItemManager;
     using UI_DSM.Shared.DTO.Common;
     using UI_DSM.Shared.DTO.Models;
     using UI_DSM.Shared.Models;
@@ -88,6 +89,7 @@ namespace UI_DSM.Server.Modules
                 return;
             }
 
+            _ = context.RequestServices.GetService<IAnnotatableItemManager>();
             dto.Author = participant.Id;
             await base.CreateEntity(manager, dto, context, deepLevel);
         }
@@ -109,6 +111,18 @@ namespace UI_DSM.Server.Modules
                 return new RequestResponseDto();
             }
 
+            var annotation = (await manager.GetEntity(entityId)).OfType<Annotation>().FirstOrDefault();
+
+            if (annotation?.Author.Id != participant.Id)
+            {
+                context.Response.StatusCode = 403;
+
+                return new RequestResponseDto()
+                {
+                   Errors = new List<string>{"Unable to delete a Comment from someelse"}
+                };
+            }
+
             return await base.DeleteEntity(manager, entityId, context);
         }
 
@@ -128,6 +142,15 @@ namespace UI_DSM.Server.Modules
 
             if (participant == null)
             {
+                return;
+            }
+
+            _ = context.RequestServices.GetService<IAnnotatableItemManager>();
+            var annotation = (await manager.GetEntity(entityId)).OfType<Annotation>().FirstOrDefault();
+
+            if (annotation?.Author.Id != participant.Id && annotation?.Content != dto.Content)
+            {
+                context.Response.StatusCode = 403;
                 return;
             }
